minor changes, bug fixes, debug logging changes

+++ iodine-api/src/main/java/hu/trigary/iodine/api/gui/element/base/GuiElement.java
@@ -17,7 +17,7 @@ public interface GuiElement<T extends GuiElement<T>> extends AttachmentHolder, D
 	/**
 	 * The inclusive upper bound for padding values.
 	 */
-	int PADDING_UPPER_BOUND = Byte.MAX_VALUE;
+	int PADDING_UPPER_BOUND = Short.MAX_VALUE;
 	
 	/**
 	 * Gets the GUI that contains this element.
